Refactor: Move DataService.BroadCastDataResult one level up
